requirements.txt
datasets==2.17.1
evaluate==0.4.2
langchain_core==0.1.52
matplotlib==3.7.5
networkx==2.8.8
numpy==1.24.3
numpy==1.24.4
pandas==2.0.3
peft==0.10.0
pykan==0.0.3
rdflib==7.0.0
scikit_learn==1.3.1
Scrapy==2.11.1
seaborn==0.13.2
tensorboard==2.13.0
tiktoken==0.7.0
torch==2.3.0
tqdm==4.66.1
transformers==4.38.1
trl==0.8.6
validators==0.22.0
